requirements.txt
flask>=2.2.2
python-dotenv>=1.0.0
openai>=0.27.8